+++ pallets/proofs/src/post/mod.rs
@@ -5,9 +5,9 @@ use frame_support::{pallet_prelude::*, sp_runtime::BoundedBTreeMap};
 use polka_storage_proofs::get_partitions_for_window_post;
 use primitives::{
     commitment::RawCommitment,
-    proofs::{PublicReplicaInfo, RegisteredPoStProof, Ticket, MAX_SECTORS_PER_PROOF},
+    proofs::{PublicReplicaInfo, RegisteredPoStProof, Ticket},
     sector::SectorNumber,
-    NODE_SIZE,
+    MAX_SECTORS_PER_PROOF, NODE_SIZE,
 };
 use sha2::{Digest, Sha256};
