+++ pallets/storage-provider/src/proofs.rs
@@ -3,7 +3,7 @@ use frame_support::{
     pallet_prelude::{ConstU32, RuntimeDebug},
     sp_runtime::BoundedVec,
 };
-use primitives::proofs::{RegisteredPoStProof, MAX_POST_PROOF_BYTES};
+use primitives::{proofs::RegisteredPoStProof, MAX_POST_PROOF_BYTES};
 use scale_info::TypeInfo;
 use sp_core::blake2_64;
