+++ storagext/cli/src/cmd/market.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use anyhow::bail;
 use clap::{ArgGroup, Subcommand};
-use primitives::proofs::DealId;
+use primitives::DealId;
 use storagext::{
     deser::DeserializablePath,
     multipair::{DebugPair, MultiPairSigner},
